feat: add GeneralBleScanner class for Bluetooth device scanning and integrate it into the project
